Move org member table to its own component in OrgMembers.tsx
